pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillsctl"
version = "0.1.0"
description = "Multi-repo prompt/skill cleanup & migration tool"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.17",
    "click>=8.1",
    "requests>=2.31",
    "gitpython>=3.1",
    "rich>=13.0",
    "rank-bm25>=0.2",
]

[project.optional-dependencies]
embeddings = [
    "openai>=1.0",
    "numpy>=1.24",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "responses>=0.24",
]

[project.scripts]
skillsctl = "skillsctl.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skillsctl*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
